+++ src/ports/web/api/tests.rs
@@ -1,15 +1,14 @@
 use super::*;
-
-use crate::{adapters::json, core::usecases, test::Bencher};
-
-use crate::core::util::sort::Rated;
+use crate::{adapters::json, core::usecases, core::util::sort::Rated, test::Bencher};
 
 pub mod prelude {
-    pub use crate::core::db::*;
-    use crate::infrastructure::db::{sqlite, tantivy};
-    pub use crate::infrastructure::flows::prelude as flows;
-    pub use crate::ports::web::tests::prelude::*;
-    use crate::ports::web::{self, api};
+    pub use crate::{
+        core::db::*, infrastructure::flows::prelude as flows, ports::web::tests::prelude::*,
+    };
+    use crate::{
+        infrastructure::db::{sqlite, tantivy},
+        ports::web::{self, api},
+    };
 
     pub fn setup() -> (Client, sqlite::Connections) {
         let (client, conn, _) = web::tests::setup(vec![("/", api::routes())]);
@@ -1719,4 +1718,26 @@ fn entries_export_csv() {
     let req = client.get("/export/entries.csv?bbox=-1,-1,1,1");
     let response = req.dispatch();
     assert_eq!(response.status(), Status::Unauthorized);
+}
+
+#[test]
+fn check_place_duplicate() {
+    let (client, db) = setup();
+    let res = client.post("/entries")
+                    .header(ContentType::JSON)
+                    .body(r#"{"title":"foo","description":"bla","lat":0.0,"lng":0.0,"categories":["x"],"license":"CC0-1.0","tags":[]}"#)
+                    .dispatch();
+    assert_eq!(res.status(), Status::Ok);
+    let mut res = client.post("/duplicates/check-place")
+                    .header(ContentType::JSON)
+                    .body(r#"{"title":"fooO","description":"bla","lat":0.001,"lng":0.0,"categories":["y"],"license":"CC0-1.0","tags":[]}"#)
+                    .dispatch();
+    assert_eq!(res.status(), Status::Ok);
+    test_json(&res);
+    let body_str = res.body().and_then(|b| b.into_string()).unwrap();
+    let eid = db.exclusive().unwrap().all_places().unwrap()[0]
+        .0
+        .id
+        .clone();
+    assert_eq!(body_str, format!("[[\"{}\", \"SimilarChars\"]]", eid));
 }